Refactor nearcache to its own internal package
